chore(api): move Probe handlers from beta to v2 package
