+++ src/main/java/io/cryostat/net/web/http/api/v2/TargetProbePostHandler.java
@@ -35,7 +35,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.cryostat.net.web.http.api.beta;
+package io.cryostat.net.web.http.api.v2;
 
 import java.util.Map;
 import java.util.Set;
@@ -43,6 +43,7 @@ import java.util.Set;
 import javax.inject.Inject;
 
 import io.cryostat.core.agent.AgentJMXHelper;
+import io.cryostat.core.agent.LocalProbeTemplateService;
 import io.cryostat.core.log.Logger;
 import io.cryostat.core.sys.Environment;
 import io.cryostat.core.sys.FileSystem;
@@ -52,30 +53,54 @@ import io.cryostat.net.TargetConnectionManager;
 import io.cryostat.net.security.ResourceAction;
 import io.cryostat.net.web.http.HttpMimeType;
 import io.cryostat.net.web.http.api.ApiVersion;
-import io.cryostat.net.web.http.api.v2.AbstractV2RequestHandler;
-import io.cryostat.net.web.http.api.v2.IntermediateResponse;
-import io.cryostat.net.web.http.api.v2.RequestParameters;
 
 import com.google.gson.Gson;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.ext.web.handler.impl.HttpStatusException;
 import org.apache.commons.lang3.StringUtils;
 
-public class TargetProbeDeleteHandler extends AbstractV2RequestHandler<Void> {
+/**
+ * TargetProbePostHandler will facilitate adding probes to a target and will have the following form
+ * and response types:
+ *
+ * <p>POST /api/v2/targets/:targetId/probes/
+ *
+ * <p>targetId - The location of the target JVM to connect to, in the form of a service:rmi:jmx://
+ * JMX Service URL. Should use percent-encoding.
+ *
+ * <p>Parameters
+ *
+ * <p>probeTemplate - name of the probe template to use
+ *
+ * <p>Responses
+ *
+ * <p>200 - No body
+ *
+ * <p>401 - User authentication failed. The body is an error message. There will be an
+ * X-WWW-Authenticate: $SCHEME header that indicates the authentication scheme that is used.
+ *
+ * <p>404 - The target could not be found. The body is an error message.
+ *
+ * <p>427 - JMX authentication failed. The body is an error message. There will be an
+ * X-JMX-Authenticate: $SCHEME header that indicates the authentication scheme that is used.
+ */
+public class TargetProbePostHandler extends AbstractV2RequestHandler<Void> {
 
-    static final String PATH = "targets/:targetId/probes";
+    static final String PATH = "targets/:targetId/probes/:probeTemplate";
 
     private final Logger logger;
     private final NotificationFactory notificationFactory;
+    private final LocalProbeTemplateService probeTemplateService;
     private final FileSystem fs;
     private final TargetConnectionManager connectionManager;
     private final Environment env;
-    private static final String NOTIFICATION_CATEGORY = "ProbeTemplateDeleted";
+    private static final String NOTIFICATION_CATEGORY = "ProbeTemplateUploaded";
 
     @Inject
-    TargetProbeDeleteHandler(
+    TargetProbePostHandler(
             Logger logger,
             NotificationFactory notificationFactory,
+            LocalProbeTemplateService service,
             FileSystem fs,
             AuthManager auth,
             TargetConnectionManager connectionManager,
@@ -84,6 +109,7 @@ public class TargetProbeDeleteHandler extends AbstractV2RequestHandler<Void> {
         super(auth, gson);
         this.logger = logger;
         this.notificationFactory = notificationFactory;
+        this.probeTemplateService = service;
         this.connectionManager = connectionManager;
         this.env = env;
         this.fs = fs;
@@ -101,7 +127,7 @@ public class TargetProbeDeleteHandler extends AbstractV2RequestHandler<Void> {
 
     @Override
     public HttpMethod httpMethod() {
-        return HttpMethod.DELETE;
+        return HttpMethod.POST;
     }
 
     @Override
@@ -118,33 +144,35 @@ public class TargetProbeDeleteHandler extends AbstractV2RequestHandler<Void> {
     public IntermediateResponse<Void> handle(RequestParameters requestParams) throws Exception {
         Map<String, String> pathParams = requestParams.getPathParams();
         String targetId = pathParams.get("targetId");
-        StringBuilder sb = new StringBuilder();
-        if (StringUtils.isBlank(targetId)) {
-            sb.append("targetId is required.");
+        String probeTemplate = pathParams.get("probeTemplate");
+        if (StringUtils.isAnyBlank(targetId, probeTemplate)) {
+            StringBuilder sb = new StringBuilder();
+            if (StringUtils.isBlank(targetId)) {
+                sb.append("targetId is required.");
+            }
+            if (StringUtils.isBlank(probeTemplate)) {
+                sb.append("\"probeTemplate\" is required.");
+            }
             throw new HttpStatusException(400, sb.toString().trim());
         }
-        try {
-            return connectionManager.executeConnectedTask(
-                    getConnectionDescriptorFromParams(requestParams),
-                    connection -> {
-                        connection.connect();
-                        AgentJMXHelper helper = new AgentJMXHelper(connection.getHandle());
-                        // The convention for removing probes in the agent controller mbean is to
-                        // call
-                        // defineEventProbes with a null argument.
-                        helper.defineEventProbes(null);
-                        notificationFactory
-                                .createBuilder()
-                                .metaCategory(NOTIFICATION_CATEGORY)
-                                .metaType(HttpMimeType.JSON)
-                                .message(Map.of("targetId", targetId))
-                                .build()
-                                .send();
-                        return new IntermediateResponse<Void>().body(null);
-                    });
-        } catch (Exception e) {
-            throw e;
-        }
+        return connectionManager.executeConnectedTask(
+                getConnectionDescriptorFromParams(requestParams),
+                connection -> {
+                    connection.connect();
+                    AgentJMXHelper helper = new AgentJMXHelper(connection.getHandle());
+                    helper.defineEventProbes(probeTemplateService.getTemplate(probeTemplate));
+                    notificationFactory
+                            .createBuilder()
+                            .metaCategory(NOTIFICATION_CATEGORY)
+                            .metaType(HttpMimeType.JSON)
+                            .message(
+                                    Map.of(
+                                            Map.of("targetId", targetId),
+                                            Map.of("probeTemplate", probeTemplate)))
+                            .build()
+                            .send();
+                    return new IntermediateResponse<Void>().body(null);
+                });
     }
 
     @Override
