+++ src/test/java/io/cryostat/net/web/http/api/v2/TargetProbeDeleteHandlerTest.java
@@ -35,7 +35,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.cryostat.net.web.http.api.beta;
+package io.cryostat.net.web.http.api.v2;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.lenient;
@@ -48,7 +48,6 @@ import javax.management.ObjectName;
 import org.openjdk.jmc.rjmx.IConnectionHandle;
 
 import io.cryostat.MainModule;
-import io.cryostat.core.agent.LocalProbeTemplateService;
 import io.cryostat.core.log.Logger;
 import io.cryostat.core.net.JFRConnection;
 import io.cryostat.core.sys.Environment;
@@ -61,8 +60,6 @@ import io.cryostat.net.TargetConnectionManager;
 import io.cryostat.net.security.ResourceAction;
 import io.cryostat.net.web.http.HttpMimeType;
 import io.cryostat.net.web.http.api.ApiVersion;
-import io.cryostat.net.web.http.api.v2.IntermediateResponse;
-import io.cryostat.net.web.http.api.v2.RequestParameters;
 
 import com.google.gson.Gson;
 import io.vertx.core.MultiMap;
@@ -79,11 +76,10 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-public class TargetProbeGetHandlerTest {
+public class TargetProbeDeleteHandlerTest {
 
-    TargetProbesGetHandler handler;
+    TargetProbeDeleteHandler handler;
     @Mock AuthManager auth;
-    @Mock LocalProbeTemplateService templateService;
     @Mock FileSystem fs;
     @Mock Logger logger;
     @Mock NotificationFactory notificationFactory;
@@ -108,19 +104,19 @@ public class TargetProbeGetHandlerTest {
         lenient().when(notificationBuilder.message(Mockito.any())).thenReturn(notificationBuilder);
         lenient().when(notificationBuilder.build()).thenReturn(notification);
         this.handler =
-                new TargetProbesGetHandler(
-                        auth, targetConnectionManager, notificationFactory, gson);
+                new TargetProbeDeleteHandler(
+                        logger, notificationFactory, fs, auth, targetConnectionManager, env, gson);
     }
 
     @Nested
     class BasicHandlerDefinition {
         @Test
-        void shouldBeGETHandler() {
-            MatcherAssert.assertThat(handler.httpMethod(), Matchers.equalTo(HttpMethod.GET));
+        void shouldBeDELETEHandler() {
+            MatcherAssert.assertThat(handler.httpMethod(), Matchers.equalTo(HttpMethod.DELETE));
         }
 
         @Test
-        void shouldBeV2API() {
+        void shouldBeBetaAPI() {
             MatcherAssert.assertThat(handler.apiVersion(), Matchers.equalTo(ApiVersion.V2));
         }
 
@@ -137,8 +133,8 @@ public class TargetProbeGetHandlerTest {
         }
 
         @Test
-        void shouldReturnJSONMimeType() {
-            MatcherAssert.assertThat(handler.mimeType(), Matchers.equalTo(HttpMimeType.JSON));
+        void shouldReturnPlaintextMimeType() {
+            MatcherAssert.assertThat(handler.mimeType(), Matchers.equalTo(HttpMimeType.PLAINTEXT));
         }
 
         @Test
@@ -151,6 +147,9 @@ public class TargetProbeGetHandlerTest {
     class Requests {
 
         @Mock RequestParameters requestParams;
+        private static final String AGENT_OBJECT_NAME =
+                "org.openjdk.jmc.jfr.agent:type=AgentController";
+        private static final String DEFINE_EVENT_PROBES = "defineEventProbes";
 
         @Test
         public void shouldRespondOK() throws Exception {
@@ -169,15 +168,14 @@ public class TargetProbeGetHandlerTest {
                                             .execute(connection));
             Mockito.when(connection.getHandle()).thenReturn(handle);
             Mockito.when(handle.getServiceOrDummy(MBeanServerConnection.class)).thenReturn(mbsc);
-            Object result = Mockito.mock(Object.class);
             Mockito.when(
                             mbsc.invoke(
                                     any(ObjectName.class),
                                     any(String.class),
                                     any(Object[].class),
                                     any(String[].class)))
-                    .thenReturn(result);
-            IntermediateResponse<String> response = handler.handle(requestParams);
+                    .thenReturn(null);
+            IntermediateResponse<Void> response = handler.handle(requestParams);
             MatcherAssert.assertThat(response.getStatusCode(), Matchers.equalTo(200));
         }
 
@@ -185,7 +183,7 @@ public class TargetProbeGetHandlerTest {
         public void shouldRespond400WhenTargetIdIsMissing() throws Exception {
             Mockito.when(requestParams.getPathParams()).thenReturn(Map.of("targetId", ""));
             try {
-                IntermediateResponse<String> response = handler.handle(requestParams);
+                IntermediateResponse<Void> response = handler.handle(requestParams);
             } catch (HttpStatusException e) {
                 MatcherAssert.assertThat(e.getStatusCode(), Matchers.equalTo(400));
             }
